Added three todo list files (#4)
I did the project in a separate repository so I had to move it here. I
also didn't use shadow DOM, so the JavaScript isn't encapsulated, but it
could be very easily manipulated to include a template and the
shadowDOM.

I better understood it by doing it without ShadowDOM. Although after
talking in class, I could very well utilize that if need be.

I also filled it with comments so it's easy to understand the process I
took.